Use native f32 math constants instead of casting from f64
The exp() and log() methods were casting std::f64::consts values to f32
with `as`, which required #[allow(clippy::cast_possible_truncation)]
annotations to keep clippy quiet. Rust's standard library provides
these same constants as native f32 values in std::f32::consts, so we
can reference them directly and drop the cast and its suppression.

+++ src/tensor.rs
@@ -694,17 +694,15 @@ impl Tensor {
 
     /// Natural exponential implemented as `2^(x * log2(e))`.
     #[must_use]
-    #[allow(clippy::cast_possible_truncation)]
     pub fn exp(&self) -> Self {
-        let log2e = Self::scalar(std::f64::consts::LOG2_E as f32, self.device());
+        let log2e = Self::scalar(std::f32::consts::LOG2_E, self.device());
         self.mul(&log2e).exp2()
     }
 
     /// Natural logarithm implemented as `log2(x) * ln(2)`.
     #[must_use]
-    #[allow(clippy::cast_possible_truncation)]
     pub fn log(&self) -> Self {
-        let ln2 = Self::scalar(std::f64::consts::LN_2 as f32, self.device());
+        let ln2 = Self::scalar(std::f32::consts::LN_2, self.device());
         self.log2().mul(&ln2)
     }
